lab-34

+++ src/components/todo/todo.js
@@ -2,12 +2,15 @@ import React, { useState, useEffect, useContext } from 'react';
 import TodoForm from './form.js';
 import TodoList from './list.js';
 import {Container, Row, Col } from 'react-bootstrap';
-import Pagination from 'react-bootstrap-4-pagination';
+import Auth from '../auth/auth.js';
+
 
 
 
 import './todo.scss';
 import axios from 'axios';
+import { LoginContext } from '../../context/auth-context.js';
+
 
 
 
@@ -18,43 +21,26 @@ const todoAPI = 'https://api-js401.herokuapp.com/api/v1/todo';
 
 const ToDo = () => {
 
- 
-
+  const context = useContext(LoginContext);
+  
+  
+  
+  
+  
   const [list, setList] =useState([]);
   const [count, setCount] =useState(0);
   
-
+  
   const addItem = (item) => {
-    item.due = new Date();
-    console.log(JSON.stringify(item));
-    axios( {
-      url: todoAPI,
-      method: 'post',
-      mode: 'cors',
-      cache: 'no-cache',
-      headers: { 'Content-Type': 'application/json' },
-      data: JSON.stringify(item),
-    })
-      .then(response => response.data)
-      .then(savedItem => {
-        setList([...list, savedItem]);
-      })
-      .catch(console.error);
-  };
-
-  const toggleComplete = id => {
-
-    let item = list.filter(i => i._id === id)[0] || {};
-
-    if (item._id) {
-
-      item.complete = !item.complete;
-
-      let apiUrl = `${todoAPI}/${id}`;
-
+    
+    if (context.user.capabilities.includes('create')){
+    
+      console.log(context.user.capabilities.includes('create'));
+      item.due = new Date();
+      console.log(JSON.stringify(item));
       axios( {
-        url: apiUrl,
-        method: 'put',
+        url: todoAPI,
+        method: 'post',
         mode: 'cors',
         cache: 'no-cache',
         headers: { 'Content-Type': 'application/json' },
@@ -62,35 +48,73 @@ const ToDo = () => {
       })
         .then(response => response.data)
         .then(savedItem => {
-          setList(list.map(listItem => listItem._id === item._id ? savedItem : listItem));
+          setList([...list, savedItem]);
         })
         .catch(console.error);
+    } else {
+      console.error('not authorized');
+    }
+    
+    
+  };
+
+  const toggleComplete = (id, displayList) => {
+    if (context.user.capabilities.includes('update')){
+
+      let item = displayList.filter(i => i._id === id)[0] || {};
+
+      if (item._id) {
+
+        item.complete = !item.complete;
+
+        let apiUrl = `${todoAPI}/${id}`;
+
+        axios( {
+          url: apiUrl,
+          method: 'put',
+          mode: 'cors',
+          cache: 'no-cache',
+          headers: { 'Content-Type': 'application/json' },
+          data: JSON.stringify(item),
+        })
+          .then(response => response.data)
+          .then(savedItem => {
+            setList(list.map(listItem => listItem._id === item._id ? savedItem : listItem));
+          })
+          .catch(console.error);
+      }
+    }else {
+      console.error('not authorized');
     }
 
   };
 
-  const deleteItem = id => {
+  const deleteItem = (id, displayList) => {
+    if (context.user.capabilities.includes('delete')){
 
-    let item = list.filter(i => i._id === id)[0] || {};
+      let item = displayList.filter(i => i._id === id)[0] || {};
 
-    if (item._id) {
+      if (item._id) {
 
 
-      let apiUrl = `${todoAPI}/${id}`;
+        let apiUrl = `${todoAPI}/${id}`;
 
-      axios( {
-        url: apiUrl,
-        method: 'delete',
-        mode: 'cors',
-        cache: 'no-cache',
-        headers: { 'Content-Type': 'application/json' },
-        data: JSON.stringify(item),
-      })
-       
-        .then(response => {
-          getTodoItems();
+        axios( {
+          url: apiUrl,
+          method: 'delete',
+          mode: 'cors',
+          cache: 'no-cache',
+          headers: { 'Content-Type': 'application/json' },
+          data: JSON.stringify(item),
         })
-        .catch(console.error);
+       
+          .then(response => {            
+            getTodoItems();            
+          })
+          .catch(console.error);
+      }
+    }else {
+      console.error('not authorized');
     }
 
   };
@@ -108,6 +132,8 @@ const ToDo = () => {
 
   useEffect(getTodoItems, []);
 
+  useEffect(getTodoItems, [context.loggedIn]);
+
   useEffect(() => {
     let number = list.filter(item => !item.complete).length;
     setCount(number);    
@@ -118,34 +144,36 @@ const ToDo = () => {
 
   return (
     <>
-      <Container fluid>
-        <Row fluid>
-          <section className="count-message">
-            <h2>
+      <Auth capability="read">
+        <Container fluid>
+          <Row fluid>
+            <section className="count-message">
+              <h2>
             There are {count} items to complete.
-            </h2>
-          </section>
-        </Row>
+              </h2>
+            </section>
+          </Row>
         
-        <Row>
-          <section className="todo">
-            <Col>
-              <div>
-                <TodoForm handleSubmit={addItem} />
-              </div>
-            </Col>
-            <Col>
-              <div>
-                <TodoList
-                  list={list}
-                  handleComplete={toggleComplete}
-                  handleDelete={deleteItem}
-                />
-              </div>
-            </Col>
-          </section>
-        </Row>
-      </Container>
+          <Row>
+            <section className="todo">
+              <Col>
+                <div>
+                  <TodoForm handleSubmit={addItem} />
+                </div>
+              </Col>
+              <Col>
+                <div>
+                  <TodoList
+                    list={list}
+                    handleComplete={toggleComplete}
+                    handleDelete={deleteItem}
+                  />
+                </div>
+              </Col>
+            </section>
+          </Row>
+        </Container>
+      </Auth>
     </>
   );
 };
